on peut changer la taille de la note

+++ UI-rusty-synth/Mase.Ui/Models/NoteModel.cs
@@ -5,6 +5,6 @@ public class NoteModel
     public string Id       { get; set; } = System.Guid.NewGuid().ToString();
     public int    Pitch    { get; set; }
     public int    Start    { get; set; }
-    public int    Length   { get; set; } = 1;
+    public double Length   { get; set; } = 1.0;
     public double Velocity { get; set; } = 0.9;
 }
